modularizing endpoint types restful server works with docker
