+++ docker-compose.yml
@@ -1,6 +1,6 @@
 web:
   build: .
-  command: python -u server.py
+  command: python -u restful_server.py
   ports:
     - "5000:5000"
   volumes:
