+++ requirements.txt
@@ -2,3 +2,4 @@ Flask==0.10.1
 protobuf==3.0.0b3
 grpcio==0.13.1
 pymongo==3.2.2
+Flask-RESTful==0.3.5
